Use long alignment for scratch buffer
Fixes last of scratch alignment warnings in hb-coretext.

+++ src/hb-buffer-private.hh
@@ -186,7 +186,7 @@ struct hb_buffer_t {
   HB_INTERNAL bool make_room_for (unsigned int num_in, unsigned int num_out);
   HB_INTERNAL bool shift_forward (unsigned int count);
 
-  typedef int scratch_buffer_t;
+  typedef long scratch_buffer_t;
   HB_INTERNAL scratch_buffer_t *get_scratch_buffer (unsigned int *size);
 
   inline void clear_context (unsigned int side) { context_len[side] = 0; }
